refactor: Move search function to Contact model
